+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).All notable changes to this project will be documented in this file.
 
-## Unreleased
+## 1.3.0 - 01/01/2025
 
 ### Added
 
@@ -25,6 +25,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Bytes.split_last`
 - `Bytes.remove`
 - `Bytes.swap_remove`
+- `Bytes.put_char`
+- `Bytes.extend`
 - `Vec.swap`
 - `Vec.swap_unchecked`
 - `Vec.fill`